% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCitations.R, R/getCitationsPkg.R
\docType{package}
\name{getCitations}
\alias{getCitations}
\title{getCitations}
\usage{
getCitations(Rmd, outpath, bibpath = NULL)
}
\arguments{
\item{Rmd}{the rmarkdown file to parse}

\item{outpath}{the location of the new citations file}

\item{bibpath}{specifcy the master citations file path manuals (defaults to current bibliography in the Rmd file)}
}
\description{
See README on \href{https://github.com/RichardJActon/getCitations/}{GitHub}
}
